Initial implementation of Tube joining
Join Tubes based on the results from the flowSOM clustering. This will
cross-join cells from selected tubes for joining.

This will produce a large number of additional data, how much this will
impact classification performance will need to be evaluated. Using a
20x20 node matrix will generate about 10x the number of events per file.
Subsampling for FlowSOM generation will probably need to be used to
retain usability for larger datasets.

Currently metadata is not correctly saved, thus the information on
channel range is lost on saving the data to fcs files. Either this needs
to be implemented correctly or the data needs to be serialized in
another way.