Added saliency plot generation without maximization
Previously saliency plots were generated with the maximum of the
gradients over all channels. The changes made to the keras vis package
now allow for plotting of saliency values for specific channels. There
are multiple functions which were added of which the following three are
the most important:

1. calc_salinecy
With this function the gradients/saliency values for a specific case,
model and class can be computed. The gradients can then be used in
different plotting functions.
2. draw_saliency_heatmap
This function plots a heatmap of the saliency values for the sommap
nodes of the given case and tube. The gradients passed into the function
obviously have to be computed for the same case. The list that is
returned contains figures for each class.
3. plot_tube
This can be considered as the main plotting function. For each of the 2D
combinations of channels colors corresponding to the maximum of the
saliency values of the two channels are picked and then a scatterplot is
created. This requires a mapping from somnodes back to the events. The
function returns as many figures as the amount of classes.

Additionally, I added an updated example on how to generate plots for a
specific case and model.